requirements.txt
# Filepath: e:\capstone 2024\CAPSTONE REBIRTH\Build_From_Scratch\VID\vid-backend\requirements.txt
# --- Core AI & LangChain Libraries ---
google-generativeai==0.5.4
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-google-genai==1.0.2

# --- Vector Store & Document Processing ---
faiss-cpu==1.8.0
PyMuPDF==1.24.1
pandas==2.2.2
numpy==1.26.4
watchdog==4.0.1
# --- Web App & API ---
streamlit==1.35.0
fastapi==0.111.0
uvicorn[standard]==0.29.0

# --- Database ---
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9

# --- Utilities ---
python-dotenv==1.0.1